Attempting to extract all of bluetooth to one class.
