+++ app/src/main/java/net/kwatts/powtools/App.java
@@ -2,6 +2,7 @@ package net.kwatts.powtools;
 
 
 import android.app.Application;
+import android.os.PowerManager;
 
 import net.kwatts.powtools.util.SharedPreferencesUtil;
 
@@ -12,6 +13,8 @@ public class App extends Application {
 
     public static App INSTANCE = null;
     private SharedPreferencesUtil sharedPreferencesUtil = null;
+    PowerManager.WakeLock mWakeLock;
+
 
     public App() {
         INSTANCE = this;
@@ -22,5 +25,26 @@ public class App extends Application {
             sharedPreferencesUtil = new SharedPreferencesUtil(App.this);
         }
         return sharedPreferencesUtil;
+    }
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+
+        initWakeLock();
+    }
+
+    private void initWakeLock() {
+        PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
+        assert powerManager != null;
+        mWakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "pOWToolsWakeLock");
+    }
+
+    public void acquireWakeLock() {
+        mWakeLock.acquire();
+    }
+
+    public void releaseWakeLock() {
+        mWakeLock.release();
     }
 }
