+++ app/src/main/java/net/kwatts/powtools/util/Util.java
@@ -29,4 +29,74 @@ public class Util {
         }
         return null;
     }
+
+    /* Helper methods. For dealing with bytes see:
+       http://www.roseindia.net/java/master-java/bitwise-bitshift-operators.shtml
+       https://calleerlandsson.com/2014/02/06/rubys-bitwise-operators/
+     */
+    public static short byteToShort(byte[] v) {
+        return (short)((v[1] << 8) + (v[0] & 0xff));
+    }
+
+    // 1 byte/8-bit signed two's complement (-128 to 127)
+    // returns an int, 32-bit signed two's complement
+    public static int unsignedByte(byte var0) {
+        return var0 & 255;
+    }
+
+    // short = 2 bytes/16-bit signed two's complement (-32,768 to 32,767)
+    public static int unsignedShort(byte[] var0) {
+        int var1;
+        if(var0.length < 2) {
+            var1 = -1;
+        } else {
+            var1 = (unsignedByte(var0[0]) << 8) + unsignedByte(var0[1]);
+        }
+        // ByteBuffer.wrap(v_bytes).getShort() also works
+        // or
+        // ByteBuffer bb = ByteBuffer.allocate(2);
+        // bb.order(ByteOrder.LITTLE_ENDIAN);
+        // bb.put(var0[0]);
+        // bb.put(var0[1]);
+        // short shortVal = bb.getShort(0);
+        return var1;
+    }
+
+    // double is 64 bit and used for decimals
+    public static double cel2far(int celsius) {
+        return (9.0/5.0)*celsius + 32;
+    }
+    public static double far2cel(int far) {
+        return (5.0/9.0)*(far - 32);
+    }
+    public static double milesToKilometers(double paramDouble) {
+        return paramDouble * 1.609344;
+
+    }
+    public static double revolutionsToKilometers(double paramDouble)
+    {
+        return paramDouble * 35.0D / 39370.099999999999D;
+    }
+
+    public static double revolutionsToMiles(double paramDouble)
+    {
+        return paramDouble * 35.0D / 63360.0D;
+    }
+    public static double rpmToKilometersPerHour(double paramDouble)
+    {
+        return 60.0D * (35.0D * paramDouble) / 39370.099999999999D;
+    }
+
+    public static double rpmToMilesPerHour(double paramDouble)
+    {
+        return 60.0D * (35.0D * paramDouble) / 63360.0D;
+    }
+
+    public static String bytesToHex(byte c[]) {
+        StringBuilder sb = new StringBuilder();
+        for (byte b : c) {
+            sb.append(String.format("%02x", b));
+        }
+        return sb.toString();
+    }
 }
